Create Simple Linear Regression Model without Modules
This is a Simple Linear Regression model I created without any modules. The principle is that the program takes the average of the slopes and y-intercepts of each line connecting each point to the next, and takes the average, creating another line. When using this please keep in mind that you must not include anything in coords_list other than numbers. For example, if the coordinates are (0,2), and (2,4), then it must be entered as [0,2,2,4]. Thank you!